+++ examples/blockly-react/src/Blockly/BlocklyComponent.jsx
@@ -21,52 +21,51 @@
  * @author samelh@google.com (Sam El-Husseini)
  */
 
- import React from 'react';
- import './BlocklyComponent.css';
- import {useEffect, useRef} from 'react';
+import React from 'react';
+import './BlocklyComponent.css';
+import {useEffect, useRef} from 'react';
 
- import Blockly from 'blockly/core';
- import {javascriptGenerator} from 'blockly/javascript';
- import locale from 'blockly/msg/en';
- import 'blockly/blocks';
+import Blockly from 'blockly/core';
+import {javascriptGenerator} from 'blockly/javascript';
+import locale from 'blockly/msg/en';
+import 'blockly/blocks';
 
- Blockly.setLocale(locale);
+Blockly.setLocale(locale);
 
- function BlocklyComponent(props) {
-    const blocklyDiv = useRef();
-    const toolbox = useRef();
-    let primaryWorkspace = useRef();
+function BlocklyComponent(props) {
+  const blocklyDiv = useRef();
+  const toolbox = useRef();
+  let primaryWorkspace = useRef();
 
-    const generateCode = () => {
-        var code = javascriptGenerator.workspaceToCode(
-          primaryWorkspace.current
-        );
-        console.log(code);
+  const generateCode = () => {
+    var code = javascriptGenerator.workspaceToCode(primaryWorkspace.current);
+    console.log(code);
+  };
+
+  useEffect(() => {
+    const {initialXml, children, ...rest} = props;
+    primaryWorkspace.current = Blockly.inject(blocklyDiv.current, {
+      toolbox: toolbox.current,
+      ...rest,
+    });
+
+    if (initialXml) {
+      Blockly.Xml.domToWorkspace(
+        Blockly.utils.xml.textToDom(initialXml),
+        primaryWorkspace.current,
+      );
     }
+  }, [primaryWorkspace, toolbox, blocklyDiv, props]);
 
-    useEffect(() => {
-        const { initialXml, children, ...rest } = props;
-            primaryWorkspace.current = Blockly.inject(
-                blocklyDiv.current,
-                {
-                    toolbox: toolbox.current,
-                    ...rest
-                },
-            );
-
-            if (initialXml) {
-                Blockly.Xml.domToWorkspace(Blockly.utils.xml.textToDom(initialXml), primaryWorkspace.current);
-            }
-    }, [primaryWorkspace, toolbox, blocklyDiv, props]);
-
-    return (
+  return (
     <React.Fragment>
-        <button onClick={generateCode}>Convert</button>
-        <div ref={blocklyDiv} id="blocklyDiv" />
-        <div style={{ display: 'none' }} ref={toolbox}>
-            {props.children}
-        </div>
-    </React.Fragment>);
+      <button onClick={generateCode}>Convert</button>
+      <div ref={blocklyDiv} id="blocklyDiv" />
+      <div style={{display: 'none'}} ref={toolbox}>
+        {props.children}
+      </div>
+    </React.Fragment>
+  );
 }
 
 export default BlocklyComponent;