+++ examples/blockly-vue3/src/main.js
@@ -4,9 +4,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { createApp } from "vue";
-import App from "./App.vue";
+import {createApp} from 'vue';
+import App from './App.vue';
 
 const app = createApp(App);
 
-app.mount("#app");
+app.mount('#app');