+++ examples/custom-generator-codelab/src/toolbox.js
@@ -5,35 +5,35 @@
  */
 
 export const toolbox = {
-  'kind': 'flyoutToolbox',
-  'contents': [
+  kind: 'flyoutToolbox',
+  contents: [
     {
-      'kind': 'block',
-      'type': 'object',
+      kind: 'block',
+      type: 'object',
     },
     {
-      'kind': 'block',
-      'type': 'member',
+      kind: 'block',
+      type: 'member',
     },
     {
-      'kind': 'block',
-      'type': 'math_number',
+      kind: 'block',
+      type: 'math_number',
     },
     {
-      'kind': 'block',
-      'type': 'text',
+      kind: 'block',
+      type: 'text',
     },
     {
-      'kind': 'block',
-      'type': 'logic_boolean',
+      kind: 'block',
+      type: 'logic_boolean',
     },
     {
-      'kind': 'block',
-      'type': 'logic_null',
+      kind: 'block',
+      type: 'logic_null',
     },
     {
-      'kind': 'block',
-      'type': 'lists_create_with',
+      kind: 'block',
+      type: 'lists_create_with',
     },
   ],
 };