+++ plugins/renderer-inline-row-separators/test/.mocharc.js
@@ -1,6 +1,0 @@
-'use strict';
-
-module.exports = {
-  ui: 'tdd',
-  require: ['jsdom-global/register'],
-};